Beginning high-level overhaul and restructuring
